refactor: 공지 아이템 컴포넌트 만듦
