--- src/components/common/notice/Notice.css
@@ -1,35 +1,0 @@
-.notice-item {
-  /* background-color: aqua; */
-  line-height: 2;
-}
-
-.notice-item .title,
-.notice-item .description,
-.notice-item .edit-date {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  padding: var(--space-xsmall);
-}
-
-.notice-item .title,
-.notice-item .description {
-  border-bottom: 1px solid var(--color-light-gray);
-  font-size: var(--font-medium);
-}
-
-.notice-item .title {
-  font-weight: 700;
-}
-
-.notice-item .edit-date {
-  font-size: var(--font-small);
-}
-
-.notice-item img {
-  width: 100%;
-  height: 200px;
-  object-fit: cover;
-  border-top-left-radius: var(--base-border-radius);
-  border-top-right-radius: var(--base-border-radius);
-}
